+++ dwds/lib/dwds.dart
@@ -103,7 +103,8 @@ class Dwds {
     bool useSseForDebugBackend,
     bool serveDevTools,
     UrlEncoder urlEncoder,
-    bool spawnDds = true,
+    bool spawnDds,
+    bool enableDevtoolsLaunch,
   }) async {
     hostname ??= 'localhost';
     enableDebugging ??= true;
@@ -111,6 +112,8 @@ class Dwds {
     useSseForDebugProxy ??= true;
     useSseForDebugBackend ??= true;
     serveDevTools ??= true;
+    enableDevtoolsLaunch ??= true;
+    spawnDds ??= true;
     globalLoadStrategy = loadStrategy;
 
     DevTools devTools;
@@ -146,6 +149,7 @@ class Dwds {
     var injected = DwdsInjector(
       loadStrategy,
       extensionUri: extensionUri,
+      enableDevtoolsLaunch: enableDevtoolsLaunch,
     );
 
     var devHandler = DevHandler(